+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## master
 
+* Enhancements
+  * Cache environment config in application config (#393)
+
+* Bug Fixes
+  * fix request url port in payloads for HTTPS requests  (#391)
+
+* Breaking Changes
+  * Change default `included_environments` to only include `:prod` by default (#370)
+  * Change default event send type to :none instead of :async (#341)
+  * Make hackney an optional dependency, and simplify Sentry.HTTPClient behaviour (#400)
+
 ## 7.2.4 (2020-03-09)
 
 * Enhancements