+++ CHANGELOG.md
@@ -17,8 +17,9 @@
   * Move excluded exceptions from Sentry.Plug to Sentry.DefaultEventFilter (#402)
   * Remove Sentry.Plug and Sentry.Phoenix.Endpoint in favor of Sentry.PlugContext and Sentry.PlugCapture (#402)
   * Remove feedback form rendering and configuration (#402)
-  * Logger metadata is now specified by key in LoggerBackend instead of enabled/disabled
-  * `Sentry.capture_exception/1` now only accepts exceptions
+  * Logger metadata is now specified by key in LoggerBackend instead of enabled/disabled (#403)
+  * Require Elixir 1.10 and optionally plug_cowboy 2.3 (#403)
+  * `Sentry.capture_exception/1` now only accepts exceptions (#403)
 
 ## 7.2.4 (2020-03-09)